Created lib.php and define test_input only once. Fixes #1
